src/test/java/com/ds/springbootaws/controller/HomeControllerTest.java
package com.ds.springbootaws.controller;

import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.test.web.servlet.MockMvc;


@WebMvcTest(controllers = HomeController.class)
public class HomeControllerTest {

  @Autowired
  private MockMvc mockMvc;

  @Test
  @WithMockUser(roles="USER")
  public void Hello가_리턴된다() throws Exception {
    String hello = "hello";

    mockMvc.perform(get("/hello"))
        .andExpect(status().isOk())
        .andExpect(content().string(hello));
  }

  @Test
  @WithMockUser(roles="USER")
  @DisplayName("hellloDto가 리턴된다")
  public void returnHelloDto() throws Exception {
    String name = "helloDto";
    int amount = 1000;

    mockMvc.perform(get("/hello/dto")
        .param("name", name)
        .param("amount", String.valueOf(amount)))
        .andExpect(status().isOk())
        .andExpect(jsonPath("$.name").value(name))
        .andExpect(jsonPath("$.amount").value(amount));
  }
}
